Rename result.rs to layout.rs
