+++ src/lib.rs
@@ -9,10 +9,10 @@ extern crate alloc;
 extern crate serde;
 
 pub mod geometry;
+pub mod layout;
 pub mod node;
 pub mod number;
 pub mod prelude;
-pub mod result;
 pub mod style;
 
 mod flexbox;
