+++ src/prelude.rs
@@ -1,8 +1,8 @@
 pub use crate::{
     geometry::{Rect, Size},
+    layout::Layout,
     node::{Node, Sprawl},
     number::Number,
-    result::Layout,
     style::{
         AlignContent, AlignItems, AlignSelf, Dimension, Direction, Display, FlexDirection, FlexWrap, JustifyContent,
         Overflow, PositionType, Style,
